+++ x/btcstkconsumer/types/query.pb.go
@@ -640,7 +640,7 @@ func (m *FinalityProviderResponse) GetConsumerId() string {
 	return ""
 }
 
-// QueryFinalityProviderConsumerRequest requests the consumer id a BSN finality
+// QueryFinalityProviderConsumerRequest requests the consumer id a Consumer finality
 // provider belongs to
 type QueryFinalityProviderConsumerRequest struct {
 	// fp_btc_pk_hex is the hex str of Bitcoin secp256k1 PK of the finality provider
@@ -687,7 +687,7 @@ func (m *QueryFinalityProviderConsumerRequest) GetFpBtcPkHex() string {
 	return ""
 }
 
-// QueryFinalityProviderConsumerResponse returns the BSN finality provider
+// QueryFinalityProviderConsumerResponse returns the Consumer finality provider
 // consumer id
 type QueryFinalityProviderConsumerResponse struct {
 	ConsumerId string `protobuf:"bytes,1,opt,name=consumer_id,json=consumerId,proto3" json:"consumer_id,omitempty"`