+++ x/zoneconcierge/keeper/ibc_packet_btc_headers.go
@@ -24,11 +24,11 @@ func (k Keeper) BroadcastBTCHeaders(ctx context.Context) error {
 	//   - If last segment valid: Returns headers from last sent to tip
 	//   - If last segment invalid (reorg): Returns last w+1 headers from tip
 	//
-	// TODO: Should use BSN base BTC header as starting point:
-	// - If no headers sent: Return from BSN base to tip
+	// TODO: Should use Consumer base BTC header as starting point:
+	// - If no headers sent: Return from Consumer base to tip
 	// - If headers previously sent:
 	//   - If last segment valid: Return from last sent header to tip
-	//   - If last segment invalid (reorg): Return from BSN base to tip
+	//   - If last segment invalid (reorg): Return from Consumer base to tip
 	headers := k.getHeadersToBroadcast(ctx)
 	if len(headers) == 0 {
 		k.Logger(sdkCtx).Info("skipping BTC header broadcast",